requirements.txt
discord.py[voice] 
python-dotenv 
youtube_dl 
PyNaCl
yt_dlp
youtube-search-python
lyricsgenius
pytube